Move DownloadContainer and DownloadView tests into single Download test file.
