+++ src/uniprotkb/components/__mocks__/mockApi.ts
@@ -8,6 +8,9 @@ import mockFasta from './fasta.json';
 
 const mock = new MockAdapter(axios);
 mock.onGet(/.+noresult/).reply(200, noResults, { 'x-totalrecords': 0 });
+mock
+  .onGet(/\/uniprotkb\/search.*&format=fasta.*/)
+  .reply(200, mockFasta, { 'content-type': 'text/fasta' });
 mock.onGet(/\/uniprotkb\/search/).reply(200, results, { 'x-totalrecords': 25 });
 mock.onGet(/\/uniprotkb\/result-fields/).reply(200, resultFields);
 mock
